+++ src/azure-cli/azure/cli/command_modules/vm/tests/latest/test_vm_commands.py
@@ -1416,15 +1416,19 @@ class VMCreateExistingOptions(ScenarioTest):
         self.cmd('vm show -n {vm} -g {rg}',
                  checks=self.check('storageProfile.osDisk.vhd.uri', 'https://{sa}.blob.core.windows.net/{container}/{disk}.vhd'))
 
-    @ResourceGroupPreparer(name_prefix='cli_test_vm_create_vmss_')
-    def test_vm_create_vmss(self, resource_group):
+    @ResourceGroupPreparer(name_prefix='cli_test_vm_reference_vmss_')
+    def test_vm_reference_vmss(self, resource_group):
         self.kwargs.update({
-            'vmss': 'vmss1',
-            'vm': 'vm1'
+            'vm': 'vm1',
+            'vmss': 'vmss1'
         })
 
-        self.cmd('vmss create -g {rg} -n {vmss} --image UbuntuLTS')
-        self.cmd('vm create -g {rg} -n {vm} --image UbuntuLTS --vmss {vmss}')
+        self.cmd('vmss create -g {rg} -n {vmss} --empty')
+        self.cmd('vm create -g {rg} -n {vm} --image ubuntults --vmss {vmss}')
+        vmss_id = self.cmd('vmss show -g {rg} -n {vmss}').get_output_in_json()['id']
+        self.cmd('vm show -g {rg} -n {vm}', checks=[
+            self.check('virtualMachineScaleSet.id', vmss_id)
+        ])
 
     @ResourceGroupPreparer(name_prefix='cli_test_vm_create_provision_vm_agent_')
     def test_vm_create_provision_vm_agent(self, resource_group):